# RestBuilder-Demo/RestBuilder-Demo-impl/rest-openapi.yaml
info:
    description:
        "RestBuilderDemo REST API. A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
        'com.liferay.assignment.restbuilder.client', and version '1.0.0'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "RestBuilderDemo"
    version: v1.0
openapi: 3.0.1

components:
  schemas:
    Entity:
      description: A very simple entity
      properties:
        name:
          description: The entity name.
          type: string
        id:
          description: The entity ID.
          type: integer
      type: object
      
paths:
  "/entities/{entityId}":
    get:
      operationId: getEntity
      parameters:
        - in: path
          name: entityId
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
            application/xml:
              schema:
                $ref: "#/components/schemas/Entity"
          description: ""
      tags: ["Entity"]
      
    delete:
      operationId: deleteEntity
      parameters:
        - in: path
          name: entityId
          required: true
          schema:
            type: integer
      responses:
        204:
          content:
            application/json: {}
            application/xml: {}
          description: ""
      tags: ["Entity"]
      
    put:
      operationId: putEntity
      parameters:
        - in: path
          name: entityId
          required: true
          schema:
              type: integer
              
      requestBody:
        content:
            application/json:
              schema:
                  $ref: "#/components/schemas/Entity"
                    
            application/xml:
              schema:
                  $ref: "#/components/schemas/Entity"
                    
      responses:
          200:
              content:
                  application/json:
                      schema:
                          $ref: "#/components/schemas/Entity"
                  application/xml:
                      schema:
                          $ref: "#/components/schemas/Entity"
              description: ""
      tags: ["Entity"]
      
  "/entities":
    get:
      operationId: getEntitiesPage
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                 $ref: "#/components/schemas/Entity"
                type: array
            application/xml:
              schema:
                items:
                 $ref: "#/components/schemas/Entity"
                type: array
          description: ""
      tags: ["Entity"]
      
    post:
      operationId: postEntity
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Entity"
          application/xml:
            schema:
              $ref: "#/components/schemas/Entity"

      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
            application/xml:
              schema:
                $ref: "#/components/schemas/Entity"
          description: ""
      tags: ["Entity"]
    
  
   
    
      
      
   